# 구현 | state mgr 마이그레이션

+++ Project/Scripts/CBackgroundScript.cpp
@@ -12,7 +12,7 @@ void CBackgroundScript::LoadFromFile(ifstream& _File)
 }
 
 CBackgroundScript::CBackgroundScript()
-	: CScript(SCRIPT_TYPE::BACKGROUNDSCRIPT)
+	: CScript((UINT)SCRIPT_TYPE::BACKGROUNDSCRIPT)
 {
 }
